fix bug mes paris

+++ backend/routes/bets.js
@@ -13,6 +13,7 @@ router.post(
     const bet = new Bet({
       title: req.body.title,
       content: req.body.content,
+      creator: req.userData.userId
     });
     bet.save().then(createdBet => {
       res.status(201).json({
@@ -36,7 +37,7 @@ router.put(
       content: req.body.content,
       userId: req.userData.userId
     });
-    Bet.updateOne({ _id: req.params.id}, bet).then(result => {
+    Bet.updateOne({ _id: req.params.id, creator: req.userData.userId }, bet).then(result => {
       if (result.nModified > 0){
         res.status(200).json({ message: 'update'});
       } else {
@@ -66,7 +67,7 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Bet.deleteOne({ _id: req.params.id}).then(result => {
+  Bet.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
     if (result.n > 0){
       res.status(200).json({ message: 'Bet deleted'});
     } else {
